requester: Add new API to support multiple transports
This patch has two goals: (1) enable consumers to send PLDM messages
over different transports and (2) do this in a way that allows us to
move towards a cleaner and more complete requester API.

The sole transport option of MCTP via the userspace -demux-daemon is
being deprecated. New transports are being added: MCTP via the kernel
(AF_MCTP) and eventually NC-SI. The requester API needs updating to
support multiple transports, as well as not have MCTP specific details
in the API. To avoid a flag day, the current API has been rewritten
terms of the new API.

The current APIs operate at the transport level - they don't implement
all of the behaviour necessary for a requester. As such, the new API
functions to send/recv a message have the prefix `pldm_transport`,
rather than `pldm_requester`. Given the level that these API are
operating at, these only send and receive a PLDM message. Any additional
logic such as looking for a response with a particular instance ID
belongs at the requester abstraction level.

Some of the missing behaviours to fully be a PLDM requester are:
assigning instance IDs, request retransmission, implementing timeouts,
and enforcing only one PLDM request to a specific TID at a time. These
things are currently implemented in pldmd, meaning any consumer other
than pldmd using the libpldm "requester" API doesn't get the
functionality of a requester and has to implement these things
themselves.

We would like to eventually move these behaviours into libpldm so the
libpldm requester API is actually implmenting what is required to be a
PLDM requester.

This patch starts the move of instance id handling into libpldm.
Currently pldmd allocates instance ids, and exposes a DBus call for
external apps to get instance ids. It relies on exploiting a behaviour
of mctp-demux to reclaim used ids by monitoring all incoming PLDM
responses. When moving to AF_MCTP, PLDM messages are not broadcast to
all PLDM sockets, and so this method of reclaiming ids won't work
automatically.

As instance id handling is going into libpldm eventually, we may as well
do that now rather than adding an additional step to address issue. So
an allocate and free function for instance ids is added in this patch.
As this implementation uses file locking, we have the feature of any ids
belonging to a process that dies being matically reclaimed into the id
pool.

The next steps to add in a full requester API, while enabling the use of
multiple transports looks something like this:

1) convert pldmd to use libpldm instance id API - so all users of PLDM
   are still using the same instance id allocation method.
2) convert all consumers of libpldm over to using the new libpldm API,
   including the instance id functions.
3) add in the AF_MCTP transport and move consumers over to it.
4) refactor the encode/decode functions to only encode/decode and not
   frame the message (ie, remove the instance id from these functions)
5) add additional requester functionality into libpldm, and have these
   use the `pldm_transport` API directly
6) move consumers over to the new `pldm_requester` API
7) remove unused code from pldmd

Signed-off-by: Rashmica Gupta <rashmica@linux.ibm.com> Change-Id:
I06e602831f360bbd0efda53d410bfb5080b3100d